+++ test/mongodb/teardown.js
@@ -1,4 +1,0 @@
-use dart_orm_test;
-
-db.runCommand( { dropAllUsersFromDatabase: 1, writeConcern: { w: "majority" } } );
-db.dropDatabase();